+++ app/build.gradle
@@ -42,6 +42,8 @@ kapt {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
+
+    // support
     implementation "com.android.support:appcompat-v7:$supportLib"
     implementation "com.android.support:support-annotations:${supportLib}"
     implementation "com.android.support:recyclerview-v7:${supportLib}"
@@ -78,7 +80,6 @@ dependencies {
     implementation "com.squareup.okhttp3:okhttp:$okhttpVersion"
     implementation "com.squareup.okhttp3:logging-interceptor:$okhttpVersion"
     implementation "com.jakewharton.retrofit:retrofit2-rxjava2-adapter:${rxJavaAdapter}"
-    implementation 'com.3sidedcube.util:GeoGson:1.6'
 
     //rxJava
     implementation "io.reactivex.rxjava2:rxjava:${rxJavaVersion}"
