+++ app/src/main/java/com/hack/kind/lvivbikemap/presentation/menu/AboutFragment.kt
@@ -1,4 +1,4 @@
-package com.hack.kind.lvivbikemap
+package com.hack.kind.lvivbikemap.presentation.menu
 
 import android.content.Intent
 import android.net.Uri
@@ -7,6 +7,7 @@ import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.hack.kind.lvivbikemap.R
 import kotlinx.android.synthetic.main.fragment_about.*
