+++ app/src/main/java/com/hack/kind/lvivbikemap/presentation/menu/FeedbackFragment.kt
@@ -1,10 +1,11 @@
-package com.hack.kind.lvivbikemap
+package com.hack.kind.lvivbikemap.presentation.menu
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.hack.kind.lvivbikemap.R
 import com.hack.kind.lvivbikemap.data.api.FeedbackRequest
 import kotlinx.android.synthetic.main.fragment_feedback.*
 
@@ -14,7 +15,7 @@ class FeedbackFragment : Fragment() {
         return inflater.inflate(R.layout.fragment_feedback, container, false)
     }
 
-    private lateinit var listener: FeedbackFragment.FeedbackSendListener
+    private lateinit var listener: FeedbackSendListener
 
     interface FeedbackSendListener {
         fun onFeedbackSend(feedback: FeedbackRequest)
